update README with better bootstrapping approach
Using bash with `-c` and command substitution `$()` means we don't have to persist
anything in the /tmp directory. I initially tried this:

  url='https://api.github.com/repos/ivanbrennan/ctags-config/contents/base-install.sh' \
  bash -c "$(curl -fssl -h 'Accept: application/vnd.github-blob.raw' $url) \
       base-install.sh --github-host github.com

but the prepended url variable wasn't present in the `$()` substitution. Setting it in a
separate command fixes that, and in order to avoid polluting the user's shell, I'm running
the whole thing in a subshell.