quote zeroth arg to bash -c
Bash, when given the `-c` flag, treats the next argument as a script commands
to execute, and any following arguments as positional parameters to that script,
starting with $0. I'm passing the script name (base-install.sh) as the zeroth
arg in order to produce a sensible "Usage" error message, but you could easily
mistake it for a separate command. Hopefully, quoting it makes that mistake a
little harder to make.