smarter config checking + error avoidance
If we're going to use `set -eu` in this script to throw an error if an unset variable is
referenced and exit early when something goes wrong, provide a null fallback value when
expanding a potentially unset variable rather than toggling the `eu` set options.

The first change, `${GIT_TEMPLATE_DIR-}`, uses a null fallback value to avoid blowing up
if GIT_TEMPLATE_DIR is unset.

The second change appends a null fallback, via `|| :`, to the query for an existing git
config templateDir.