portability overload
This started out as a small change in which I changed the color variables to use ANSI-C
quoting,
  col_red=$'\e[31;01m'

This makes the output correctly colored without needing to pass the `-e` flag to echo, and
echo is somewhat inconsistent in it's behavior and optional flags across different
implementations. But this $'string' quoting style is not (yet) part of the POSIX standard,
so not all shells support it. Bash, Zsh, and Korn shells do, but Fish does not.

This got me thinking about whether I should be using echo at all. The most portable
solution seems to be using printf in conjunction with the %b format specifier, which
causes backslash escape sequences to be interpretted.

This is probably all overkill, but I don't mind getting a little more practice using
printf anyway.